+++ CHANGELOG.md
@@ -2,6 +2,7 @@ HEAD
 ----
 - `destroy` method will destroy a single instance of the plugin
 - `clear` method added
+- `onClear` and `onDestroy` callbacks added
 
 1.0.2
 -----